feat: make the device link autoload when open
