+++ vl-main/src/events/mod.rs
@@ -1,0 +1,2 @@
+pub mod module_event;
+pub mod screen_event;
