+++ vl-main/src/lib.rs
@@ -3,6 +3,8 @@
 #![feature(core_intrinsics)]
 
 use bevy::prelude::*;
+pub mod events;
+pub mod manager;
 pub mod modules;
 pub mod ui;
