+++ vl-main/src/main.rs
@@ -5,6 +5,8 @@
 #![feature(core_intrinsics)]
 
 use bevy::prelude::bevy_main;
+pub mod events;
+pub mod manager;
 pub mod modules;
 pub mod ui;
