+++ vl-main/src/manager.rs
@@ -1,0 +1,3 @@
+pub trait Manager {
+    fn modify_app(&mut self, app: &mut bevy::app::App);
+}
