+++ vl-main/src/modules/base/i_module.rs
@@ -1,14 +1,14 @@
+use busrt::async_trait;
 use futures::future::BoxFuture;
 use std::any::type_name;
 use std::fmt::Debug;
 
+#[async_trait]
 pub trait IModule: Debug + Send + Sync {
     fn is_started(&self) -> bool;
 
     // Initialize a module
-    fn start(
-        &mut self,
-    ) -> BoxFuture<Result<(), Box<dyn std::error::Error>>>;
+    async fn start(&mut self) -> anyhow::Result<()>;
 
     fn get_screen_name(&self) -> &'static str {
         let full_name = type_name::<Self>();
