+++ vl-main/src/modules/mod.rs
@@ -1,5 +1,4 @@
 pub mod base;
 #[cfg(target_os = "linux")]
 pub mod linux;
-pub mod module_event;
 pub mod module_manager;
