+++ vl-main/src/ui/mod.rs
@@ -1,3 +1,4 @@
 pub mod bevy_app;
+pub mod screen_manager;
 pub mod screens;
 pub mod virtual_keyboard;
